Better ENV variable handling and

+++ simreporter/builder.sh
@@ -5,9 +5,12 @@ while true; do
         echo "`date`: Lua tool simreporter already up to date"
     else
         ./buildrelease.py -n
-        docker exec --user=root nmeaproxy bash -c 'rm /tmp/firmware/*.zip'
-        image=`ls -rt build | tail -1`
-        docker cp ${image} nmeaproxy:/tmp/firmware
+        result=$?
+        if [ ${result} == 0 ] ; then
+            docker exec --user=root nmeaproxy bash -c 'rm /tmp/firmware/*.zip'
+            image=`ls -rt build | tail -1`
+            docker cp ${image} nmeaproxy:/tmp/firmware
+        fi
     fi
     echo "`date`: Sleeping for 60 seconds"
     sleep 60
